Split error types out of Ops module
This should still be split out to better follow the hierarchy of the
modules, as this is still very monolithic. It's at least clearer when it
is split out.
