+++ cardano-cli/src/Cardano/CLI/Shelley/Run.hs
@@ -7,7 +7,7 @@ import           Cardano.Prelude
 
 import           Control.Monad.Trans.Except (ExceptT)
 
-import           Cardano.CLI.Ops (CliError (..))
+import           Cardano.CLI.Errors (CliError(..))
 import           Cardano.CLI.Shelley.Parsers
 
 import           Cardano.CLI.Shelley.Run.Address
