+++ cardano-cli/src/Cardano/CLI/Shelley/Run/StakeAddress.hs
@@ -13,7 +13,7 @@ import           Cardano.Api (StakingVerificationKey (..),
                    shelleyRegisterStakingAddress, writeCertificate)
 import           Shelley.Spec.Ledger.Keys (hashKey)
 
-import           Cardano.CLI.Ops (CliError (..))
+import           Cardano.CLI.Errors (CliError(..))
 import           Cardano.CLI.Shelley.Parsers
