+++ cardano-cli/src/Cardano/CLI/Shelley/Run/TextView.hs
@@ -4,7 +4,8 @@ module Cardano.CLI.Shelley.Run.TextView
 
 import           Cardano.Prelude
 
-import           Cardano.CLI.Ops (CliError (..), pPrintCBOR)
+import           Cardano.CLI.Errors (CliError(..))
+import           Cardano.CLI.Ops (pPrintCBOR)
 import           Cardano.CLI.Shelley.Parsers
 
 import           Cardano.Config.TextView
