+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Transaction.hs
@@ -12,7 +12,7 @@ import           Cardano.Config.Shelley.ColdKeys
                    (KeyType(..), KeyRole(..), KeyError(..), renderKeyType)
 import           Cardano.Config.TextView
 import           Cardano.CLI.Environment (readEnvSocketPath)
-import           Cardano.CLI.Ops (CliError (..))
+import           Cardano.CLI.Errors (CliError(..))
 
 import           Cardano.Config.Types
